move all connector infrastructure into a single directory
