+++ src/hackingBuddyGPT/capabilities/psexec_run_command.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 from typing import Tuple
 
-from hackingBuddyGPT.utils import PSExecConnection
+from hackingBuddyGPT.utils.connectors.psexec import PSExecConnection
 
 from ..capability import Capability
