+++ src/hackingBuddyGPT/usecases/minimal_linux_privesc.py
@@ -2,7 +2,8 @@ from typing import List
 from hackingBuddyGPT.capabilities import SSHRunCommand, SSHTestCredential
 from hackingBuddyGPT.usecases.base import use_case
 from hackingBuddyGPT.strategies import CommandStrategy
-from hackingBuddyGPT.utils import SSHConnection, llm_util
+from hackingBuddyGPT.utils import llm_util
+from hackingBuddyGPT.utils.connectors.ssh_connection import SSHConnection
 
 from mako.template import Template
