+++ src/hackingBuddyGPT/utils/__init__.py
@@ -3,6 +3,4 @@ from .console import *
 from .db_storage import *
 from .llm_util import *
 from .openai import *
-from .psexec import *
-from .ssh_connection import *
 from .ui import *
