--- src/hackingBuddyGPT/utils/local_shell/__init__.py
@@ -1,3 +1,0 @@
-from .local_shell import LocalShellConnection
-
-__all__ = ["LocalShellConnection"]
