--- src/hackingBuddyGPT/utils/psexec/__init__.py
@@ -1,3 +1,0 @@
-from .psexec import PSExecConnection
-
-__all__ = ["PSExecConnection"]
