--- src/hackingBuddyGPT/utils/ssh_connection/__init__.py
@@ -1,3 +1,0 @@
-from .ssh_connection import SSHConnection
-
-__all__ = ["SSHConnection"]
